Feat: Añadido controler de lienzo y generacion de bytes

+++ Backend/src/main/kotlin/org/pin/backend/dto/PointDeltaDTO.kt
@@ -1,3 +1,3 @@
 package org.pin.backend.dto
 
-data class PointDeltaDTO (val x: Float, val y: Float, val size: Float, val color: Int)
+data class PointDeltaDTO (val x: Short, val y: Short, val size: Float, val color: Byte)
